Moved grountruth functions
